wtp/webrtc/RTCTrackEvent-constructor.html

<!doctype html>
<meta charset="utf-8">
<title>RTCTrackEvent constructor</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  'use strict';
  
  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html
  
  /*
    5.7.  RTCTrackEvent
      [Constructor(DOMString type, RTCTrackEventInit eventInitDict)]
      interface RTCTrackEvent : Event {
        readonly attribute RTCRtpReceiver           receiver;
        readonly attribute MediaStreamTrack         track;
        [SameObject]
        readonly attribute FrozenArray<MediaStream> streams;
        readonly attribute RTCRtpTransceiver        transceiver;
      };
       dictionary RTCTrackEventInit : EventInit {
        required RTCRtpReceiver        receiver;
        required MediaStreamTrack      track;
                 sequence<MediaStream> streams = [];
        required RTCRtpTransceiver     transceiver;
      };
   */
  
  test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var receiver = transceiver.receiver;
    var track = receiver.track;
  
  
    var trackEvent = new RTCTrackEvent('track', {
      receiver: receiver, track: track, transceiver: transceiver
    });
  
    assert_equals(trackEvent.receiver, receiver);
    assert_equals(trackEvent.track, track);
    assert_array_equals(trackEvent.streams, []);
    assert_equals(trackEvent.streams, trackEvent.streams); // [SameObject]
    assert_equals(trackEvent.transceiver, transceiver);
  
    assert_equals(trackEvent.type, 'track');
    assert_false(trackEvent.bubbles);
    assert_false(trackEvent.cancelable);
  }, 'new RTCTrackEvent() with valid receiver, track, transceiver should succeed');
  
  test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var receiver = transceiver.receiver;
    var track = receiver.track;
  
  
    var stream = new MediaStream([track]);
  
    var trackEvent = new RTCTrackEvent('track', {
      receiver: receiver, track: track, transceiver: transceiver,
      streams: [stream]
    });
  
    assert_equals(trackEvent.receiver, receiver);
    assert_equals(trackEvent.track, track);
    assert_array_equals(trackEvent.streams, [stream]);
    assert_equals(trackEvent.transceiver, transceiver);
  }, 'new RTCTrackEvent() with valid receiver, track, streams, transceiver should succeed');
  
  test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var receiver = transceiver.receiver;
    var track = receiver.track;
  
  
    var stream1 = new MediaStream([track]);
    var stream2 = new MediaStream([track]);
  
    var trackEvent = new RTCTrackEvent('track', {
      receiver: receiver, track: track, transceiver: transceiver,
      streams: [stream1, stream2]
    });
  
    assert_equals(trackEvent.receiver, receiver);
    assert_equals(trackEvent.track, track);
    assert_array_equals(trackEvent.streams, [stream1, stream2]);
    assert_equals(trackEvent.transceiver, transceiver);
  }, 'new RTCTrackEvent() with valid receiver, track, multiple streams, transceiver should succeed');
  
  test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var receiver = pc.addTransceiver('audio').receiver;
    var track = pc.addTransceiver('audio').receiver.track;
  
    var stream = new MediaStream();
  
    var trackEvent = new RTCTrackEvent('track', {
      receiver: receiver, track: track, transceiver: transceiver,
      streams: [stream]
    });
  
    assert_equals(trackEvent.receiver, receiver);
    assert_equals(trackEvent.track, track);
    assert_array_equals(trackEvent.streams, [stream]);
    assert_equals(trackEvent.transceiver, transceiver);
  }, 'new RTCTrackEvent() with unrelated receiver, track, streams, transceiver should succeed');
  
  test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var receiver = transceiver.receiver;
    var track = receiver.track;
  
  
    assert_throws(new TypeError(), function () {
      return new RTCTrackEvent('track', {
        receiver: receiver, track: track
      });
    });
  }, 'new RTCTrackEvent() with no transceiver should throw TypeError');
  
  test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var receiver = transceiver.receiver;
  
  
    assert_throws(new TypeError(), function () {
      return new RTCTrackEvent('track', {
        receiver: receiver, transceiver: transceiver
      });
    });
  }, 'new RTCTrackEvent() with no track should throw TypeError');
  
  test(function (t) {
    var pc = new RTCPeerConnection();
    var transceiver = pc.addTransceiver('audio');
    var receiver = transceiver.receiver;
    var track = receiver.track;
  
  
    assert_throws(new TypeError(), function () {
      return new RTCTrackEvent('track', {
        track: track, transceiver: transceiver
      });
    });
  }, 'new RTCTrackEvent() with no receiver should throw TypeError');
  
  /*
    Coverage Report
      Interface tests are counted as 1 trivial test
       Tested         1
      Total          1
   */
</script>
